requirements.txt
streamlit
langchain
google-generativeai
langchain-google-genai
python-dotenv
firebase-admin
oauth2client
google-auth
google-cloud-firestore
google-auth-oauthlib
pycryptodome
google-cloud-bigquery 
pandas
altair
db-dtypes
langchain_community
openai
streamlit-authenticator
pyarrow
matplotlib
seaborn
scikit-learn
statsmodels
scipy
statsmodels
shap
google-cloud-storage
google-cloud-monitoring
